pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mouse>=0.7.1",
    "pillow>=12.2.0",
    "pywin32>=311",
]
